style

+++ ping-pong/thread/main.cpp
@@ -54,14 +54,11 @@ static void worker(size_t id) {
 static bool to_size_t(const char *buffer, size_t *value) {
     char *endptr;
     if (*buffer == '-') return false;
-    
     errno = 0;
-    unsigned long val = strtoul(buffer, &endptr, 0);
-    
+    size_t val = strtoul(buffer, &endptr, 0);
     if (errno == ERANGE || endptr == buffer || *endptr != '\0' || val > SIZE_MAX) {
         return false;
     }
-    
     *value = static_cast<size_t>(val);
     return true;
 }
